refactor: flatten tests
